🗑️ refactor: remove IndexingPage and clean up routing
- Deleted `IndexingPage` component to streamline the admin interface.
- Removed related routing from `TinaAdmin` for improved clarity.
